+++ CHANGELOG.md
@@ -11,6 +11,7 @@ Please view this file on the master branch, on stable branches it's out of date.
 ### Added
 - User can select segmentation run mode before start (@mkasztelnik)
 - Possibility to configure custom Ansys licenses for pipeline computation (@mkasztelnik)
+- POossibility to define scripted step deployment types (@mkasztelnik)
 
 ### Changed